[feat]: {대회 문제 목록} 관련 컴포넌트명 일괄 변경 (#109)
