+++ app/contests/[cid]/submits/components/NoneUsersContestSubmitListItem.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default function NoneContestSubmitListItem() {
+export default function NoneUsersContestSubmitListItem() {
   return (
     <tr className="border-b dark:border-gray-700 text-xs text-center">
       <th
@@ -10,6 +10,7 @@ export default function NoneContestSubmitListItem() {
         1
       </th>
       <td className="text-sm">조회된 제출 정보가 없습니다</td>
+      <td className="font-medium">-</td>
       <td className="font-medium">-</td>
       <td className="font-medium">-</td>
       <td className="font-medium">-</td>
