+++ app/contests/[cid]/submits/components/UsersContestSubmitListItem.tsx
@@ -5,7 +5,7 @@ interface ContestSubmitListItemProps {
   cid: string;
 }
 
-export default function ContestSubmitListItem({
+export default function UsersContestSubmitListItem({
   cid,
 }: ContestSubmitListItemProps) {
   const router = useRouter();
@@ -28,7 +28,7 @@ export default function ContestSubmitListItem({
       <td className="">2020123123</td>
       <td className="">홍길동</td>
       <td className="">A+B</td>
-      <td className="text-blue-500 font-semibold">정답</td>
+      <td className="text-[#0076C0] font-semibold">정답</td>
       <td>
         <span>1527 </span>
         <span className="ml-[-1px] text-red-500">KB</span>
